fix bugs in restarting from checkpoint (#31)

+++ src/exabiome/nn/train.py
@@ -39,7 +39,7 @@ def parse_args(*addl_args, argv=None):
     parser.add_argument('output', type=str, help='file to save model', default=None)
 
     add_dataset_arguments(parser)
-    parser.add_argument('-F', '--features', type=str, help='a checkpoint file for previously trained features', default=None)
+    parser.add_argument('-F', '--features_checkpoint', type=str, help='a checkpoint file for previously trained features', default=None)
 
     parser.add_argument('-l', '--load', action='store_true', default=False, help='load data into memory before running training loop')
 
@@ -128,7 +128,7 @@ def process_args(args=None, return_io=False):
 
     # make sure we are classifying if we are using adding classifier layers
     # to a resnet features model
-    if args.features is not None:
+    if args.features_checkpoint is not None:
         if args.manifold:
             raise ValueError('Cannot use manifold loss (i.e. -M) if adding classifier (i.e. -F)')
         args.classify = True
@@ -176,7 +176,9 @@ def process_args(args=None, return_io=False):
     del args.gpus
 
     if args.debug:
-        targs['fast_dev_run'] = True
+        targs['limit_train_batches'] = 10
+        targs['limit_val_batches'] =5
+        targs['max_epochs'] = 1
 
     if args.lr_find:
         targs['auto_lr_find'] = True
@@ -203,13 +205,14 @@ from pytorch_lightning import Trainer, seed_everything
 from pytorch_lightning.loggers import CSVLogger
 from pytorch_lightning.callbacks import ModelCheckpoint
 
-from mpi4py import MPI
 
-comm = MPI.COMM_WORLD
-RANK = comm.Get_rank()
+#from mpi4py import MPI
+#
+#comm = MPI.COMM_WORLD
+#RANK = comm.Get_rank()
+RANK = 0
 
 def print0(*msg, **kwargs):
-    RANK = 0
     if RANK == 0:
         print(*msg, **kwargs)
 
@@ -233,7 +236,13 @@ def run_lightning(argv=None):
             print0(f'symlinking to {args.checkpoint} from {outdir}')
             dest = output('start.ckpt')
             src = os.path.relpath(args.checkpoint, start=outdir)
-            os.symlink(src, dest)
+            if os.path.exists(dest):
+                existing_src = os.readlink(dest)
+                if existing_src != src:
+                    msg = f'Cannot create symlink to checkpoint -- {dest} already exists, but points to {existing_src}'
+                    raise RuntimeError(msg)
+            else:
+                os.symlink(src, dest)
 
     seed_everything(args.seed)
 
@@ -318,7 +327,6 @@ def overall_metric(model, loader, metric):
     return val
 
 def print_args(argv=None):
-    '''Run Lightning Learning Rate finder'''
     import pickle
     import ruamel.yaml as yaml
     parser = argparse.ArgumentParser()
@@ -328,6 +336,15 @@ def print_args(argv=None):
         namespace = pickle.load(f)
     yaml.main.safe_dump(vars(namespace), sys.stdout, default_flow_style=False)
 
+
+def show_hparams(argv=None):
+    import yaml
+    import torch
+    parser = argparse.ArgumentParser()
+    parser.add_argument('checkpoint', help='the checkpoint file to pull hparams from', type=str)
+    args = parser.parse_args(argv)
+    hparams = torch.load(args.checkpoint, map_location=torch.device('cpu'))['hyper_parameters']
+    yaml.dump(hparams, sys.stdout)
